Move Schema component specific stuff from CatalogTree to SchemaField
