+++ airbyte-webapp/src/core/domain/catalog/fieldUtil.ts
@@ -18,8 +18,6 @@ export function toInnerModel(
     throw e;
   }
 
-  result.catalog.streams.forEach((stream, i) => (stream.id = i.toString()));
-
   return result;
 }
